requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==21.6
httpx==0.27.2
celery[redis]==5.4.0
redis==5.1.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
razorpay==1.4.1
python-dateutil==2.9.0
